+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.8] - 2025-07-16
+- fix: aptos messages decoding format (#43)
+- `--sender-queue`: set tx' `--gas-limit` (if provided) and `nonce` (#44)
+
 ## [0.2.7] - 2025-05-29
 - fix: decoding of very old v1.2 messages (#33)